New bootstrap info, test refactoring (WIP)
